Upgrade `@actions/core` to `1.10.0`
This changeset upgrades `@actions/core` (and `@actions/github`
while we're at it) because of the new environment file change
that is resulting in the deprecation of previous forms of
`setOutput`.

Because this action is using `@actions/core`, upgrading the
package should be enough to fix it.

Fixes and closes bufbuild/buf-breaking-action#56.

More info:
https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/